activity-scheduler-plugin/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
google-generativeai==0.3.0
python-socketio==5.11.1
eventlet==0.35.2
